// apps/journeys-admin/__generated__/GetAdminTemplate.ts
export interface GetAdminTemplate_adminTemplate_language {
  __typename: "Language";
  id: string;
  name: GetAdminTemplate_adminTemplate_language_name[];
}
// ...
export type GetAdminTemplate_adminTemplate_blocks = GetAdminTemplate_adminTemplate_blocks_ButtonBlock | GetAdminTemplate_adminTemplate_blocks_CardBlock | GetAdminTemplate_adminTemplate_blocks_GridContainerBlock | GetAdminTemplate_adminTemplate_blocks_GridItemBlock | GetAdminTemplate_adminTemplate_blocks_IconBlock | GetAdminTemplate_adminTemplate_blocks_ImageBlock | GetAdminTemplate_adminTemplate_blocks_RadioOptionBlock | GetAdminTemplate_adminTemplate_blocks_RadioQuestionBlock | GetAdminTemplate_adminTemplate_blocks_SignUpBlock | GetAdminTemplate_adminTemplate_blocks_StepBlock | GetAdminTemplate_adminTemplate_blocks_TypographyBlock | GetAdminTemplate_adminTemplate_blocks_VideoBlock | GetAdminTemplate_adminTemplate_blocks_VideoTriggerBlock;

export interface GetAdminTemplate_adminTemplate_primaryImageBlock {
  __typename: "ImageBlock";
  id: string;
  parentBlockId: string | null;
  parentOrder: number | null;
  src: string | null;
  alt: string;
  width: number;
  height: number;
  /**
   * blurhash is a compact representation of a placeholder for an image.
   * Find a frontend implementation at https: // github.com/woltapp/blurhash
   */
  blurhash: string;
}

export interface GetAdminTemplate_adminTemplate_userJourneys_user {
  __typename: "User";
  id: string;
  firstName: string;
  lastName: string | null;
  imageUrl: string | null;
}

export interface GetAdminTemplate_adminTemplate_userJourneys {
  __typename: "UserJourney";
  id: string;
  user: GetAdminTemplate_adminTemplate_userJourneys_user | null;
}

export interface GetAdminTemplate_adminTemplate {
  __typename: "Journey";
  id: string;
  slug: string;
  title: string;
  description: string | null;
  status: JourneyStatus;
  language: GetAdminTemplate_adminTemplate_language;
  createdAt: any;
  publishedAt: any | null;
  themeName: ThemeName;
  themeMode: ThemeMode;
  seoTitle: string | null;
  seoDescription: string | null;
  template: boolean | null;
  blocks: GetAdminTemplate_adminTemplate_blocks[] | null;
  primaryImageBlock: GetAdminTemplate_adminTemplate_primaryImageBlock | null;
  userJourneys: GetAdminTemplate_adminTemplate_userJourneys[] | null;
}

export interface GetAdminTemplate {
  adminTemplate: GetAdminTemplate_adminTemplate | null;
}